Notification: Add NotificationActivity_Bongo file
